Environment Setup/templates/flask-react/backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
flask-sqlalchemy>=3.1.0
marshmallow>=3.20.0